+++ _examples/print-basic-text/main.go
@@ -1,0 +1,9 @@
+package main
+
+import "github.com/pterm/pterm"
+
+func main() {
+	pterm.DefaultBasicText.Println("Default basic text printer.")
+	pterm.DefaultBasicText.Println("Can be used in any" + pterm.LightMagenta(" TextPrinter ") + "context.")
+	pterm.DefaultBasicText.Println("For example to resolve progressbars and spinners.")
+}