+++ print.go
@@ -37,6 +37,10 @@ func Sprinto(a ...interface{}) string {
 // Spaces are added between operands when neither is a string.
 // It returns the number of bytes written and any write error encountered.
 func Print(a ...interface{}) {
+	if DisableOutput {
+		return
+	}
+
 	var ret string
 	var printed bool
 
@@ -72,6 +76,10 @@ func Printf(format string, a ...interface{}) {
 // Spaces are added between operands when neither is a string.
 // It returns the number of bytes written and any write error encountered.
 func Fprint(writer io.Writer, a ...interface{}) {
+	if DisableOutput {
+		return
+	}
+
 	color.Fprint(writer, Sprint(a...))
 }
 
@@ -90,6 +98,10 @@ func Fprintln(writer io.Writer, a ...interface{}) {
 // time.Sleep(time.Second)
 // pterm.Oprint("Hello, Earth!")
 func Printo(a ...interface{}) {
+	if DisableOutput {
+		return
+	}
+
 	color.Print("\r" + Sprint(a...))
 }