+++ pterm.go
@@ -1,1 +1,6 @@
 package pterm
+
+var (
+	// DisableOutput completely disables output from pterm. Can be used in CLI application quiet mode.
+	DisableOutput = false
+)